Exclude orders with negative prices from available funds
We and another foodcoop use the foodsoft also for deposit refunding. Therefore we have articles with negative deposit that can be ordered. These orders have a negative price and increase the available funds. This should only take place when the order is balanced. Therefore, unbalanced orders with negative price should be excluded in value_of_open_orders and value_of_finished_orders.